% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-functions.R
\name{annual_summary}
\alias{annual_summary}
\title{Summarize a brick by year}
\usage{
annual_summary(b, fun)
}
\arguments{
\item{fun}{function to apply. Default: mean.}
}
\description{
Summarize a brick by year
}
