% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation-bands.R
\name{centroids}
\alias{centroids}
\title{Centroid of polygon}
\usage{
centroids(poly_station)
}
\arguments{
\item{poly_station}{\code{\link[sf]{sf}} polygon of station catchment or a
\code{\link[raster]{extent}}.}
}
\value{
a \link[tibble:tibble-package]{tibble} with columns \code{lon} and \code{lat}.
}
\description{
Centroid of polygon
}
