% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_flow}
\alias{convert_flow}
\title{Convert between units of flow volume}
\source{
hydromad r package
}
\usage{
convert_flow(
  x,
  from = "mm",
  to = "mm",
  area.km2 = -1,
  timestep.default = "days"
)
}
\arguments{
\item{x}{a numeric vector or time series.}

\item{from}{units to convert from (see Details).}

\item{to}{units to convert into (see Details).}

\item{area.km2}{area (in square kilometres) that flow volume is averaged over.
This must be given when converting between measures of depth and measures of volume.}

\item{timestep.default}{the time step if not specified in from or to.}
}
\value{
numeric vector with flow in the new units
}
\description{
This function converts between units of flow (volumetric throughput),
designed mainly for hydrology applications. As a special case, this function
can also convert units of volume or units of depth (length).
}
\details{
This function can convert flow rates between different volume units,
or different timestep units, or both. Volume can be specified directly, or
as a depth and area combination. The unit specifications from and to can
include a time step, like "volume / timestep". A multiplier may also be
included with the time step, like "volume / N timesteps".
If no time step is specified in from and to, then it is taken to be
timestep.default.
The volume units supported are: (these can be abbreviated)
mL, cL, dL, L, daL, hL, kL, ML, GL, TL
cm^3, dm^3, m^3, km^3, ft^3
The depth units supported are: (these can be abbreviated)
mm, cm, metres, km, inches, feet
The time units supported are: (these can be abbreviated)
ms, seconds, minutes, hours, days, weeks, months, years / annum
Additionally, the value "cumecs" (cubic metres per second) is equivalent to
"m^3/sec".
}
\examples{
convert_flow(300, from = "m^3/sec", to = "mm/day", area.km2 = 10^4)
}
\author{
Felix Andrews felix@nfrac.org
}
