% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation-bands.R
\name{elev_bands_nc}
\alias{elev_bands_nc}
\title{Elevation bands NetCDF file}
\source{
\url{https://github.com/naddor/tofu/blob/master/input_output_settings/create_elev_bands_nc.R}
}
\usage{
elev_bands_nc(
  elev_tab,
  ccoords,
  file_nc = "inst/extdata/elevation_bands_74.nc",
  na = -9999,
  force_v4 = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{elev_tab}{data.frame output from \code{\link{elev_bands}}}

\item{ccoords}{a \link[tibble:tibble-package]{tibble} with columns \code{lon}
and \code{lat}.}

\item{file_nc}{character, path to NetCDF file}

\item{na}{scalar numeric, Default: -9999}
}
\value{
character, path to the NetCDF file.
}
\description{
Elevation bands NetCDF file
}
\examples{
if (FALSE) {
  elevation_tab <- elev_bands(
    con_dem = condem74, meteo_raster = precclim74, dz = 100
  )
  elev_bands_nc(
    elev_tab = elevation_tab,
    ccoords = centroids(poly74)
    file_nc = file.path(tempdir(), "elevation_bands_74.nc"),
    na = -9999
  )
}
}
