% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcdata74.R
\docType{data}
\name{forcdata74}
\alias{forcdata74}
\title{Forcing time series of hydrometeorological variables.}
\format{
A tibble with 13,149 daily observations, and 6 variables:
\describe{
\item{date}{date object, start 1980-01-01, end 2015-12-31}
\item{id}{numeric vector, id of ONS station}
\item{pr}{numeric vector, precipitation in mm/day}
\item{temp}{numeric vector, monthly average temperature in °C}
\item{pet}{numeric vector, potential ET in mm/day}
\item{q_obs}{optional, numeric vector, river discharge in mm/day}
}
}
\source{
\url{https://www.worldclim.org/data/worldclim21.html}
}
\usage{
forcdata74
}
\description{
As of Oct 2020.
}
\note{
Since FUSE require temperature only for the snow model, the daily
temperature refers to monthly climatological averages.
}
\references{
https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/joc.4518
}
\keyword{datasets}
