% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.prep-package.R
\docType{package}
\name{fuse.prep-package}
\alias{fuse.prep}
\alias{fuse.prep-package}
\title{fuse.prep: Preparation Of Input Data For Fuse}
\description{
Data tidyng of forcing time series and elevation bands for FUSE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lhmet-ped/fuse.prep}
  \item Report bugs at \url{https://github.com/lhmet-ped/fuse.prep/issues}
}

}
\author{
\strong{Maintainer}: Jônatan Tatsch \email{lhmet@ufsm.br}

}
\keyword{internal}
