% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-nc.R
\name{import_nc}
\alias{import_nc}
\title{Import the NetCDF file of daily gridded precipitation in Brazil}
\source{
\url{https://utexas.app.box.com/v/Xavier-etal-IJOC-DATA/}
}
\usage{
import_nc(varnc = c("prec", "et0"), dest_dir = "input")
}
\arguments{
\item{varnc}{a scalar character with the variable name: 'prec' or 'et0'.}

\item{dest_dir}{a character with the name of where the downloaded file is
saved. Default: \code{fusepoc-prep/input}.}
}
\value{
a character path to the NetCDF file downloaded.
}
\description{
Import the NetCDF file of daily gridded precipitation in Brazil
}
\details{
Warning: The size of the file is 1.5 GB!!! Make sure you have
space available on disk. The download of files can take a while.
}
\examples{
if(FALSE) b_prec <- fuse.prep::import_nc(varnc = "prec", dest_dir = "input")

}
\references{
https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/joc.4518
}
