% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_forcing_nc.R
\name{meteo_forcing_nc}
\alias{meteo_forcing_nc}
\title{NetCDF file of Meteorological forcings}
\usage{
meteo_forcing_nc(
  ...,
  dates,
  ccoords,
  file_nc = "inst/extdata/meteo_forcings_74.nc",
  na = -9999,
  force_v4 = TRUE
)
}
\description{
NetCDF file of Meteorological forcings
}
