% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-nc.R
\name{meteo_nc}
\alias{meteo_nc}
\title{Download NetCDF files of daily gridded precipitation and evapotranspiration}
\usage{
meteo_nc(varnc, dest_dir = "input")
}
\arguments{
\item{varnc}{a character with the variable names 'prec' and/or 'et0'.}

\item{dest_dir}{a character with the name of where the downloaded file is
saved. Default: \code{fusepoc-prep/input}.}
}
\value{
character, path to NetCDF file.
}
\description{
Download NetCDF files of daily gridded precipitation and evapotranspiration
}
\examples{
if(FALSE){
 meteo_nc("prec")
}
}
