% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-nc.R
\name{save_data}
\alias{save_data}
\title{Save data from a ONS station in a RDS file}
\usage{
save_data(
  data_posto = qnat_posto,
  .prefix = "qnat-obs-posto-",
  .posto_id = info_posto$posto[1],
  .dest_dir = "output"
)
}
\description{
Save data from a ONS station in a RDS file
}
