% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_clim}
\alias{spatial_clim}
\title{Annual Mean Climatology}
\usage{
spatial_clim(
  meteo_brick = import_nc(varnc = "prec", dest_dir = "input"),
  poly_station = poly_posto,
  fun = sum,
  save = TRUE,
  dest_dir = "output",
  ref_crs = "+proj=longlat +datum=WGS84"
)
}
\description{
Annual Mean Climatology
}
\keyword{Internal}
