% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{annual_summary}
\alias{annual_summary}
\title{Summarize a brick by year}
\usage{
annual_summary(meteo_brick, fun)
}
\arguments{
\item{meteo_brick}{\code{\link[raster]{brick}} of meteorological field
(e.g.: precipitation, evapotranspiration, etc).}

\item{fun}{function to apply. Default: mean.}
}
\description{
Summarize a brick by year
}
