% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-forcings.R
\name{comb_data}
\alias{comb_data}
\title{Join hydrometeorological data for FUSE and convert streamflow from cumecs
to mm/day.}
\usage{
comb_data(
  prec,
  et0,
  qobs,
  area,
  save = TRUE,
  prefix = "hydrodata-posto-",
  dest_dir = "output"
)
}
\arguments{
\item{prec}{a \code{\link[tibble]{tibble}} with daily precipitation (mm day^-1^)}

\item{et0}{a \code{\link[tibble]{tibble}} with daily reference evapotranspiration (mm day^-1^)}

\item{qobs}{a \code{\link[tibble]{tibble}} with daily natural streamflow (cumecs)}

\item{area}{area (in square kilometres) that flow volume is averaged
over.}

\item{save}{logical, TRUE for export output data to a RDS file.}

\item{prefix}{character, prefiz for the output file name.}

\item{dest_dir}{a character with the name of where the data is to be
saved.}
}
\value{
a \code{\link[tibble]{tibble}}, column q_obs is the \code{qnat} variable
(input) expressed in mm (discharge).
}
\description{
Join hydrometeorological data for FUSE and convert streamflow from cumecs
to mm/day.
}
