% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.prep-package.R
\docType{package}
\name{fuse.prep}
\alias{fuse.prep}
\title{\pkg{fuse.prep}: Preparation of input data for FUSE}
\description{
The goal of \pkg{fuse.prep} is to tidy and generate the input data to run FUSE
for a catchement.
}
\section{\pkg{fuse.prep} Core functions}{


\itemize{
\code{\link{elev_bands_nc}}: creates NetCDF file of elevation bands.
}

\itemize{
\code{\link{meteo_forcing_nc}}: creates NetCDF file of Meteorological
forcings.
}
}

