% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-forcings.R
\name{meteo_forcing_nc}
\alias{meteo_forcing_nc}
\title{Create NetCDF file of Meteorological forcings}
\usage{
meteo_forcing_nc(
  forc_tbl,
  ccoords,
  file_nc = "inst/extdata/74_input.nc",
  na = -9999,
  force_v4 = TRUE
)
}
\arguments{
\item{forc_tbl}{tibble with time series of meteorological forcings.}

\item{ccoords}{a \link[tibble:tibble-package]{tibble} with columns \code{lon}
and \code{lat}.}

\item{file_nc}{character, path to NetCDF file}

\item{na}{scalar numeric, Default: -9999}

\item{force_v4}{If TRUE, then the created output file will always be in netcdf-4 format (which
 supports more features, but
 cannot be read by version 3 of the netcdf library).  If FALSE, then the file is created
 in netcdf version 3 format UNLESS the user has requested features that require version 4.
 Deafult is FALSE.}

\item{dates}{objects of class "Date" representing calendar dates.}
}
\description{
Create NetCDF file of Meteorological forcings
}
\examples{
if(FALSE){
 # arquivo de saída
 forcings_nc <- "inst/extdata/posto74_input.nc"
 # exporta dados para netcdf
 meteo_forcing_nc(
   temp = forcdata74[["temp"]],
   pr = forcdata74[["pr"]],
   pet = forcdata74[["pet"]],
   q_obs = forcdata74[["q_obs"]],
   dates = forcdata74[["date"]],
   ccoords = centroids(poly_station = poly74),
   file_nc = forcings_nc
 )
 file.exists(forcings_nc)
}
}
\concept{forcings functions}
