% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_average}
\alias{spatial_average}
\title{Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station}
\usage{
spatial_average(meteo_brick, posto_poly, save = TRUE, dest_dir = "output")
}
\arguments{
\item{meteo_brick}{object of class \code{\link[brick]{raster}}}

\item{posto_poly}{bject of class \code{\link[sf]{sf}}}
}
\value{
a character path to the RDS file with a \code{\link[tibble]{tibble}}
}
\description{
Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station
}
