% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-functions.R, R/spatial.R
\name{spatial_average}
\alias{spatial_average}
\title{Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station}
\usage{
spatial_average(meteo_brick, posto_poly)

spatial_average(meteo_brick, posto_poly)
}
\arguments{
\item{meteo_brick}{object of class \code{\link[brick]{raster}}}

\item{posto_poly}{bject of class \code{\link[sf]{sf}}}

\item{poly_station}{\code{\link[sf]{sf}} polygon of station catchment or a
\code{\link[raster]{extent}}.}

\item{fun}{function to apply. Default: mean.}
}
\value{
a character path to the RDS file with a \code{\link[tibble]{tibble}}

a character path to the RDS file with a \code{\link[tibble]{tibble}}
}
\description{
Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station

Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station
}
