% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_average}
\alias{spatial_average}
\title{Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station}
\usage{
spatial_average(meteo_brick, poly_station, fun = mean)
}
\arguments{
\item{meteo_brick}{\code{\link[raster]{brick}} of meteorological field
(e.g.: precipitation, evapotranspiration, etc).}

\item{poly_station}{object of class \code{\link[sf]{sf}} and geometry
type polygon of station catchment.}

\item{fun}{function to apply. Default: mean.}
}
\value{
a character path to the RDS file with a \code{\link[tibble]{tibble}}
}
\description{
Save average precipitation or evapotranspiration over upstream drainage
area of a ONS station
}
