% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_clim}
\alias{spatial_clim}
\title{Annual Mean Climatology}
\usage{
spatial_clim(
  meteo_brick = import_nc(varnc = "prec", dest_dir = "input"),
  poly_station,
  fun = sum,
  ref_crs = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{meteo_brick}{RasterBrick of the meteorological field}

\item{poly_station}{sf polygon, station catchment}

\item{fun}{function to apply, e.g., mean and sum.}

\item{ref_crs}{character, coordinate reference system.}
}
\description{
Annual Mean Climatology
}
