% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_clim}
\alias{spatial_clim}
\title{Annual Mean Climatology}
\usage{
spatial_clim(
  meteo_brick = import_nc(varnc = "prec", dest_dir = "input"),
  poly_station,
  fun = sum,
  ref_crs = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{meteo_brick}{\code{\link[raster]{brick}} of meteorological field
(e.g.: precipitation, evapotranspiration, etc).}

\item{poly_station}{object of class \code{\link[sf]{sf}} and geometry
type polygon of station catchment.}

\item{fun}{function to apply. Default: sum.}

\item{ref_crs}{character, coordinate reference system.}
}
\description{
Annual Mean Climatology
}
